requirements.txt
aerich==0.8.1
aiofiles==24.1.0
aiosqlite==0.20.0
allure-pytest==2.13.5
allure-python-commons==2.13.5
annotated-types==0.7.0
anyio==4.7.0
asyncclick==8.1.7.2
asyncmy==0.2.10
attrs==24.3.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dictdiffer==0.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
Flask==3.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.10
iniconfig==2.0.0
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
priority==2.0.0
pydantic==2.10.4
pydantic_core==2.27.2
pypika-tortoise==0.3.2
pytest==8.3.4
pytz==2024.2
requests==2.32.3
sniffio==1.3.1
starlette==0.41.3
taskgroup==0.2.1
tomli==2.2.1
tortoise-orm==0.23.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
wsproto==1.2.0
